changed yaml format to allow ordering
also fixed config code to put config into ruby classes instead of
leaving as loose dictionaries & arrays.
items are now handled in the order they appear in the file, and the
files are handled in alphabetical order.

added support for handling an entire section at a time rather than only
as individual items.  this will be better for things like apt-get where
you can put in a whole list of what to update and do it in a single call
rather than several individual calls